Commit funcionalidade de busca das listas de chatGroups;

+++ construcao/chat/src/main/java/br/com/eits/desafio/chat/domain/service/group/MessageService.java
@@ -3,6 +3,8 @@ package br.com.eits.desafio.chat.domain.service.group;
 import java.util.Calendar;
 import java.util.List;
 
+import javax.security.auth.login.CredentialException;
+
 import org.apache.log4j.Logger;
 import org.directwebremoting.annotations.RemoteProxy;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +14,7 @@ import org.springframework.stereotype.Service;
 import br.com.eits.desafio.chat.domain.entity.group.Message;
 import br.com.eits.desafio.chat.domain.repository.group.IChatGroupRepository;
 import br.com.eits.desafio.chat.domain.repository.group.IMessageRepository;
+import br.com.eits.desafio.chat.security.ContextHolder;
 
 @Service
 @RemoteProxy (name="messageService")
@@ -41,8 +44,14 @@ public class MessageService{
 		return this.messageRepository.listAllMessagesByChatGroupId(chatGroupId);
 	}
 	
-	public void deleteMessage(Long messageId){
-		this.messageRepository.delete(messageId);
+	public void deleteMessage(Long messageId) throws CredentialException{
+		Message message = this.getMessage(messageId);
+		
+		if(message.getUserChatGroup().getUser().getId() == ContextHolder.getAuthenticatedUser().getId()){		
+			this.messageRepository.delete(messageId);
+		}else{
+			new CredentialException("User can only delete his own messages!");
+		}
 	}
 	
 	public void setMessageStatusToVisualized(Long messageId){
@@ -50,6 +59,10 @@ public class MessageService{
 		message.setVisualized(Boolean.TRUE);
 		
 		this.messageRepository.saveAndFlush(message);
+	}
+	
+	public Message getMessage(Long id){
+		return this.messageRepository.findOne(id);
 	}
